Update pip and stick to IPython 5 for py27
IPython 6.0+ does not support Python 2.7, so we need to stick to
version 5.x.  By installing a newer pip (9.0.1 at the moment),
it automatically uses the correct version of IPython.

When starting the Docker image, jupyter complained that
the `--allow-root` option must be specified, otherwise it aborts.

Minor markup fix in readme to get proper headings.